+++ feature/home/src/main/java/com/wepli/home/screen/relaylist/BannerComponent.kt
@@ -1,0 +1,33 @@
+package com.wepli.home.screen.relaylist
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.wepli.home.component.WePLiBanner
+import com.wepli.home.component.WePLiBannerType
+
+@Composable
+fun WePLiBannerLayout() {
+    val bannerList = listOf(
+        WePLiBannerType.Twitter,
+        WePLiBannerType.Instagram
+    )
+    val pagerState = rememberPagerState(pageCount = { bannerList.size })
+
+    HorizontalPager(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 16.dp),
+        state = pagerState,
+        contentPadding = PaddingValues(horizontal = 20.dp),
+        pageSpacing = 12.dp
+    ) { page ->
+        val banner = bannerList[page]
+        WePLiBanner(modifier = Modifier.fillMaxWidth(), bannerType = banner)
+    }
+}
